requirements.txt
# lua_custom_engine_generator/requirements.txt
pyyaml
tqdm